beyond_cookbook/ingredients/serializers.py
from rest_framework import serializers

from ingredients.models import Product, Unit


class ProductSerializer(serializers.ModelSerializer):

    class Meta:
        model = Product
        fields = ['name', 'description']

class UnitSerializer(serializers.ModelSerializer):

    class Meta:
        model = Unit
        fields = '__all__'