requirements.txt
Django==2.2
djangorestframework==3.10.3
psycopg2-binary==2.8.4
django-environ==0.4.5
Pillow==7.0.0